extracted almost all image manipulation methods
