+++ src/utils.js
@@ -1,6 +1,10 @@
 import { Base64 } from 'js-base64'
 import Color from 'color'
 import inside from 'point-in-polygon'
+import Trianglify from 'trianglify'
+import rough from 'roughjs/bin/wrappers/rough'
+import { Posterizer } from 'potrace'
+import Jimp from 'jimp'
 
 export function getSeed () {
   return Base64.encodeURI(String(Math.random()).substr(2))
@@ -128,9 +132,8 @@ export const getColorInBounds = (path, image, svg, debug = false) => {
 
 export const lightenPalette = (palette, amount = 0.5) => palette.map(color => {
   const parsedColor = Color(color)
-  const lightColor = parsedColor.lighten(amount).hex()
 
-  return lightColor
+  return parsedColor.lighten(amount).hex()
 })
 
 export const shuffle = (array, inplace = false) => {
@@ -144,4 +147,138 @@ export const shuffle = (array, inplace = false) => {
   }
 
   return array
+}
+
+const createThumb = (image, maxSize = 200) => {
+  let newDimensions = [maxSize, Jimp.AUTO]
+
+  if (image.getWidth() < image.getHeight()) {
+    newDimensions.reverse()
+  }
+
+  return image.clone().resize(...newDimensions)
+    .blur(2)
+}
+
+export const posterize = (image, options) => {
+  const posterizer = new Posterizer(options)
+
+  return new Promise((resolve, reject) => {
+    posterizer.loadImage(image, error => {
+      if (error) {
+        reject(error)
+      }
+
+      resolve(posterizer.getSVG())
+    })
+  })
+}
+
+const createTriangles = options =>
+  Trianglify({
+    ...options
+  }).svg({ includeNamespace: true })
+
+export function createSizedSVG (width, height) {
+  const namespaceURI = 'http://www.w3.org/2000/svg'
+  const svg = document.createElementNS(namespaceURI, 'svg')
+  svg.setAttribute('xmlns', namespaceURI)
+
+  svg.setAttribute('width', width)
+  svg.setAttribute('height', height)
+
+  return svg
+}
+
+export function extractPaths (posterizedImage) {
+  const domParser = new DOMParser()
+  const posterizedThumb = domParser.parseFromString(posterizedImage, 'image/svg+xml')
+
+  return Array.from(posterizedThumb.querySelectorAll('path'))
+}
+
+export function mergePaths (trianglePaths, posterPaths, svg) {
+  trianglePaths.forEach(path => {
+    svg.appendChild(path)
+  })
+
+  posterPaths.forEach(path => {
+    path.setAttribute('stroke', 'none')
+    svg.appendChild(path)
+  })
+
+  return svg.outerHTML
+}
+
+export function thumbnailize (image, maxSize, mime = 'image/png') {
+  const thumb = createThumb(image, maxSize)
+
+  return new Promise((resolve, reject) => {
+    thumb.getBase64(mime, (error, thumbURI) => {
+      if (error) {
+        reject(error)
+      }
+
+      resolve({
+        thumb,
+        thumbURI,
+        height: thumb.getHeight(),
+        width: thumb.getWidth()
+      })
+    })
+  })
+}
+
+export function trianglize (height, width, palette) {
+  const triangles = createTriangles({
+    height,
+    width,
+    ...(palette ? {
+      x_colors: shuffle(palette),
+      y_colors: shuffle(palette)
+    } : {})
+  })
+
+  return {
+    triangles: triangles.outerHTML,
+    trianglePaths: Array.from(triangles.querySelectorAll('path'))
+  }
+}
+
+const polygonFill = (getColor, options = {}) => ({
+  fillStyle: 'hachure',
+  roughness: 3,
+  stroke: 'transparent',
+  hachureAngle: Math.random() * 360,
+  fill: getColor(),
+  ...options
+})
+
+export function tracer (paths, thumb, width, height, simplification, getColor) {
+  const traced = createSizedSVG(width, height)
+
+  const rc = rough.svg(traced)
+
+  paths.forEach(originalPath => {
+    const polygonPath = originalPath.getAttribute('d').replace(/,\s+/g, ',')
+    const color = getColorInBounds(originalPath, thumb, traced)
+
+    const path = rc.path(polygonPath, {
+      simplification,
+      ...polygonFill(getColor, {
+        fill: `${color}`
+      })
+    })
+
+    originalPath.removeAttribute('style')
+    // originalPath.setAttribute('fill', 'none')
+    originalPath.setAttribute('fill', color)
+    originalPath.setAttribute('stroke', getColor())
+    originalPath.setAttribute('stroke', color)
+    originalPath.setAttribute('stroke-width', 1)
+    traced.appendChild(originalPath)
+    traced.appendChild(path)
+  })
+
+  return traced.outerHTML
 }
